safari-600.7-branch/LayoutTests/platform/win/fast/events/keyLocation-numpad-expected.txt

Test for bug 16410: Implement isKeypadEvent() on Windows

keydown: DOM_KEY_LOCATION_NUMPAD
keyup: DOM_KEY_LOCATION_NUMPAD
keydown: DOM_KEY_LOCATION_NUMPAD
keyup: DOM_KEY_LOCATION_NUMPAD
keydown: DOM_KEY_LOCATION_NUMPAD
keyup: DOM_KEY_LOCATION_NUMPAD
